azure-storage/requirements.txt
flixtube_common==0.*
annotated-types==0.6.0
anyio==4.3.0
azure-core==1.30.1
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
fastapi==0.110.3
h11==0.14.0
idna==3.7
isodate==0.6.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
python-multipart==0.0.9
pymongo==4.7.1
motor==3.4.0
pytest==8.2.0
python-dotenv==1.0.1
